api routes functional

+++ db/schema.sql
@@ -3,7 +3,7 @@ CREATE DATABASE smoll_talk_db;
 USE smoll_talk_db;
 
 CREATE TABLE conversation (
-  id INTEGER(11) AUTO_INCREMENT NOT NULL,
+  id INTEGER(10) AUTO_INCREMENT NOT NULL,
   topic_name VARCHAR(25),
   PRIMARY KEY (id)
 );
